pyproject.toml
[project]
name = "orcamentoofx"
version = "0.1.0"
description = "Personal Budget API. Documentation for endpoints is available at /api/docs/"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.0",
    "django>=5.2.3",
    "django-ninja>=1.4.3",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "ofxparse>=0.21",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
    "django-extensions>=4.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "pytest-subtests>=0.14.2",
    "pytest-sugar>=1.0.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.config.settings"
python_files = ["test_*.py"]
